Added buildingView to app and finished reserve logic issue
