+++ app-frontend/src/App.jsx
@@ -7,7 +7,8 @@ import Footer from './components/default-components/Footer'
 import Home from './components/home-components/Home'
 import Login from './components/login-and-signup-components/Login'
 import About from './components/home-components/About'
-import Reserve from './components/Reserve'
+import Reserve from './components/reservation-pages/Reserve'
+import BuildingView from './components/reservation-pages/Reserve'
 import SignUp from './components/login-and-signup-components/SignUp'
 import Contact from './components/home-components/Contact'
 import Forbidden from './components/default-components/Forbidden'
@@ -28,7 +29,8 @@ function App() {
           <Route path='/login' element={<RoleBasedRoute publicOnly={true} ><Login /></RoleBasedRoute>}></Route>
           <Route path='/signup' element={<RoleBasedRoute publicOnly={true}><SignUp /></RoleBasedRoute>}></Route>
           <Route path='/about' element={<About />}></Route>
-          <Route path='/reserve' element={<Reserve/>}></Route>
+          <Route path='/reserve' element={<Reserve />}></Route>
+          <Route path='/reserve:buildingId' element={<BuildingView />}></Route>
           <Route path='/contact' element={<Contact />}></Route>
           <Route path='/forbidden' element={<Forbidden />}></Route>
           <Route path='*' element={<NotFound />}></Route>
