+++ app-frontend/src/components/reservation-pages/BuildingView.jsx
@@ -1,0 +1,11 @@
+import { useState, useEffect } from "react";
+import axios from "axios";
+
+function BuildingView() {
+    <>
+        <div>
+            Hello
+        </div>
+    </>
+}
+export default BuildingView;
